Factored concurrent test runner helpers into its own file, for inclusion from multiple files
